[cudacompat] Remove CUDA-only data format classes
