+++ src/cudacompat/CUDADataFormats/SiPixelClustersSoA.h
@@ -2,7 +2,6 @@
 #define CUDADataFormats_SiPixelCluster_interface_SiPixelClustersSoA_h
 
 #include "CUDACore/cudaCompat.h"
-#include "CUDADataFormats/SiPixelClustersCUDA.h"
 
 #include <memory>
 
@@ -36,7 +35,23 @@ public:
   uint32_t const *c_moduleId() const { return moduleId_d.get(); }
   uint32_t const *c_clusModuleStart() const { return clusModuleStart_d.get(); }
 
-  using DeviceConstView = SiPixelClustersCUDA::DeviceConstView;
+  class DeviceConstView {
+  public:
+    // DeviceConstView() = default;
+
+    __device__ __forceinline__ uint32_t moduleStart(int i) const { return __ldg(moduleStart_ + i); }
+    __device__ __forceinline__ uint32_t clusInModule(int i) const { return __ldg(clusInModule_ + i); }
+    __device__ __forceinline__ uint32_t moduleId(int i) const { return __ldg(moduleId_ + i); }
+    __device__ __forceinline__ uint32_t clusModuleStart(int i) const { return __ldg(clusModuleStart_ + i); }
+
+    friend SiPixelClustersSoA;
+
+    //   private:
+    uint32_t const *moduleStart_;
+    uint32_t const *clusInModule_;
+    uint32_t const *moduleId_;
+    uint32_t const *clusModuleStart_;
+  };
 
   DeviceConstView *view() const { return view_d.get(); }
