--- src/cudacompat/CUDADataFormats/SiPixelDigiErrorsCUDA.h
@@ -1,41 +1,0 @@
-#ifndef CUDADataFormats_SiPixelDigi_interface_SiPixelDigiErrorsCUDA_h
-#define CUDADataFormats_SiPixelDigi_interface_SiPixelDigiErrorsCUDA_h
-
-#include <cuda_runtime.h>
-
-#include "CUDACore/SimpleVector.h"
-#include "CUDACore/device_unique_ptr.h"
-#include "CUDACore/host_unique_ptr.h"
-#include "DataFormats/PixelErrors.h"
-
-class SiPixelDigiErrorsCUDA {
-public:
-  SiPixelDigiErrorsCUDA() = default;
-  explicit SiPixelDigiErrorsCUDA(size_t maxFedWords, PixelFormatterErrors errors, cudaStream_t stream);
-  ~SiPixelDigiErrorsCUDA() = default;
-
-  SiPixelDigiErrorsCUDA(const SiPixelDigiErrorsCUDA&) = delete;
-  SiPixelDigiErrorsCUDA& operator=(const SiPixelDigiErrorsCUDA&) = delete;
-  SiPixelDigiErrorsCUDA(SiPixelDigiErrorsCUDA&&) = default;
-  SiPixelDigiErrorsCUDA& operator=(SiPixelDigiErrorsCUDA&&) = default;
-
-  const PixelFormatterErrors& formatterErrors() const { return formatterErrors_h; }
-
-  cms::cuda::SimpleVector<PixelErrorCompact>* error() { return error_d.get(); }
-  cms::cuda::SimpleVector<PixelErrorCompact> const* error() const { return error_d.get(); }
-  cms::cuda::SimpleVector<PixelErrorCompact> const* c_error() const { return error_d.get(); }
-
-  using HostDataError =
-      std::pair<cms::cuda::SimpleVector<PixelErrorCompact>, cms::cuda::host::unique_ptr<PixelErrorCompact[]>>;
-  HostDataError dataErrorToHostAsync(cudaStream_t stream) const;
-
-  void copyErrorToHostAsync(cudaStream_t stream);
-
-private:
-  cms::cuda::device::unique_ptr<PixelErrorCompact[]> data_d;
-  cms::cuda::device::unique_ptr<cms::cuda::SimpleVector<PixelErrorCompact>> error_d;
-  cms::cuda::host::unique_ptr<cms::cuda::SimpleVector<PixelErrorCompact>> error_h;
-  PixelFormatterErrors formatterErrors_h;
-};
-
-#endif
