--- src/cudacompat/plugin-PixelTriplets/CAHitNtupletGeneratorKernelsAlloc.cu
@@ -1,1 +1,0 @@
-#include "CAHitNtupletGeneratorKernelsAlloc.h"
