--- src/cudacompat/plugin-PixelVertexFinding/gpuVertexFinder.cu
@@ -1,1 +1,0 @@
-#include "gpuVertexFinderImpl.h"
